Patch 2024 executables to bypass resource verification
Adds code for the new April 2024 Update executable patches, which
bypass the resource verification done with ".verify" files. This makes
it possible to load mods without using devmode or needing DOOMLauncher.

The patching functions check the executable status, and if it is
detected to be an unpatched 2024 executable, it creates a backup and
applies the corresponding patch.